Began implementing compliance rate computations.
Using a script which pulls data from a view in Snowflake (which is all the EHS scores with the districts added), I successfully computed the denominator for desired division and account period.

The next step is to compute the numerators, then it will be easy to get the overall compliance rates.

Still need to figure out the best/most scalable way to get, store, and access the variables once calculated.